updated code to take larger inputs

+++ theory classes code/Prime_checker.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main()
 {
-    int num;
+    long num;
     
     printf("Enter a number : ");
-    scanf("%d", &num);
+    scanf("%ld", &num);
 
     if (num == 2)
     {
